+++ src/main/java/com/catis/model/defectresponse.java
@@ -4,20 +4,21 @@ import java.util.List;
 
 public class defectresponse {
 
-	private int inspectionid;
+	private Long inspectionid;
+
 	private List<DefectsModel> defectslist;
-	
-	public defectresponse(int inspection,List<DefectsModel> defectslist ) {
+
+	public defectresponse(Long inspection,List<DefectsModel> defectslist ) {
 		
 		this.inspectionid = inspection;
 		this.defectslist = defectslist;
 	}
 
-	public int getInspectionid() {
+	public Long getInspectionid() {
 		return inspectionid;
 	}
 
-	public void setInspectionid(int inspectionid) {
+	public void setInspectionid(Long inspectionid) {
 		this.inspectionid = inspectionid;
 	}
 
@@ -28,8 +29,4 @@ public class defectresponse {
 	public void setDefectslist(List<DefectsModel> defectslist) {
 		this.defectslist = defectslist;
 	}
-
-	
-	
-	
 }
